[Update] Idle, Move Task 하나로 통합
